Adds single-point "tusk" command line.

+++ src/vagrant.js
@@ -1,11 +1,20 @@
 var Promise = require('bluebird');
 var spawn = require('child_process').spawn;
 
-function up (cwd, next) {
+var util = require('./util');
+
+function up (cwd, opts, next) {
+  if (typeof opts == 'function') {
+    next = opts;
+    opts = {};
+  }
+  opts = opts || {};
+
   return new Promise(function (resolve, reject) {
     var p = spawn('vagrant', ['up', '--provider=google', '--no-provision'], {
       cwd: cwd,
       stdio: "inherit",
+      env: util.combine(process.env, opts.env || {}),
     });
     p.on('error', reject);
     p.on('exit', function (code) {
@@ -17,11 +26,18 @@ function up (cwd, next) {
   .nodeify(next);
 }
 
-function provision (cwd, next) {
+function provision (cwd, opts, next) {
+  if (typeof opts == 'function') {
+    next = opts;
+    opts = {};
+  }
+  opts = opts || {};
+
   return new Promise(function (resolve, reject) {
     var p = spawn('vagrant', ['provision'], {
       cwd: cwd,
       stdio: "inherit",
+      env: util.combine(process.env, opts.env || {}),
     });
     p.on('error', reject);
     p.on('exit', function (code) {
@@ -33,11 +49,18 @@ function provision (cwd, next) {
   .nodeify(next);
 }
 
-function destroy (cwd, next) {
+function destroy (cwd, opts, next) {
+  if (typeof opts == 'function') {
+    next = opts;
+    opts = {};
+  }
+  opts = opts || {};
+
   return new Promise(function (resolve, reject) {
     var p = spawn('vagrant', ['destroy', '-f'], {
       cwd: cwd,
       stdio: "inherit",
+      env: util.combine(process.env, opts.env || {}),
     });
     p.on('error', reject);
     p.on('exit', function (code) {
@@ -47,7 +70,7 @@ function destroy (cwd, next) {
     });
   })
   // TODO remove this timeout when vagrant-google waits properly
-  .delay(5000)
+  .delay(15*1000)
   .nodeify(next);
 }
